+++ Term2/Project1_Extended_Kalman_Filters/src/main.cpp
@@ -70,8 +70,8 @@ int main() {
           if (sensor_type.compare("L") == 0) {
             meas_package.sensor_type_ = MeasurementPackage::LASER;
             meas_package.raw_measurements_ = VectorXd(2);
-            float px;
-            float py;
+            double px;
+            double py;
             iss >> px;
             iss >> py;
             meas_package.raw_measurements_ << px, py;
@@ -80,9 +80,9 @@ int main() {
           } else if (sensor_type.compare("R") == 0) {
             meas_package.sensor_type_ = MeasurementPackage::RADAR;
             meas_package.raw_measurements_ = VectorXd(3);
-            float ro;
-            float theta;
-            float ro_dot;
+            double ro;
+            double theta;
+            double ro_dot;
             iss >> ro;
             iss >> theta;
             iss >> ro_dot;
@@ -91,10 +91,10 @@ int main() {
             meas_package.timestamp_ = timestamp;
           }
 
-          float x_gt;
-          float y_gt;
-          float vx_gt;
-          float vy_gt;
+          double x_gt;
+          double y_gt;
+          double vx_gt;
+          double vy_gt;
           iss >> x_gt;
           iss >> y_gt;
           iss >> vx_gt;
